♻️ refactor: refactor pluginSelectors
