+++ src/store/plugin/slices/plugin/selectors.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { PluginStoreState } from './initialState';
+import { PluginStoreState } from '../../initialState';
 import { pluginSelectors } from './selectors';
 
 const mockState = {
@@ -73,13 +73,6 @@ describe('pluginSelectors', () => {
     });
   });
 
-  describe('onlinePluginStore', () => {
-    it('should return the online plugin list', () => {
-      const result = pluginSelectors.onlinePluginStore(mockState);
-      expect(result).toEqual(mockState.pluginList);
-    });
-  });
-
   describe('pluginList', () => {
     it('should return the combined list of pluginList and customPluginList', () => {
       const result = pluginSelectors.pluginList(mockState);
@@ -126,22 +119,6 @@ describe('pluginSelectors', () => {
       };
       const result = pluginSelectors.getPluginSettingsById('plugin-1')(stateWithSettings);
       expect(result).toEqual(mockSettings);
-    });
-  });
-
-  describe('isCustomPlugin', () => {
-    it('should return false for a non-custom plugin', () => {
-      const result = pluginSelectors.isCustomPlugin('plugin-1')(mockState);
-      expect(result).toBe(false);
-    });
-
-    it('should return true for a custom plugin', () => {
-      const stateWithCustomPlugin = {
-        ...mockState,
-        customPluginList: [{ identifier: 'custom-plugin' }],
-      } as PluginStoreState;
-      const result = pluginSelectors.isCustomPlugin('custom-plugin')(stateWithCustomPlugin);
-      expect(result).toBe(true);
     });
   });
